\name{dengueoutbreak2}
\alias{dengueoutbreak2}
\docType{data}
\title{
Incidence data of the 2001/2002 single dengue outbreak in the health area "J. R. Ramirez", Havana City, Cuba.
}
\description{
This data set gives the incidence (Reported dengue cases) in each time point (in weeks)
of the 2001/2002 single dengue outbreak in the health area "J. R. Ramirez" of the Playa municipality, Havana City, as compiled by the Institute of Tropical Medicine "Pedro Kouri", Havana City, Cuba.
}
\usage{data("dengueoutbreak2")}
\format{
  A data frame with 34 observations on the following 2 variables.
  \describe{
    \item{\code{Incidence}}{a numeric vector, representing the incidence of the outbreak.}
    \item{\code{Time}}{a numeric vector, representing the time vector in weeks.}
  }
}
\source{
Institute of Tropical Medicine "Pedro Kouri", Havana City, Cuba.
}

\keyword{datasets}
