\name{DengueRT}
\alias{DengueRT}
\alias{DengueRT}
\docType{package}
\title{
Parameter Estimates and Real-Time Prediction of a Single Dengue Outbreak
}
\description{
The DengueRT package uses the incidence data from a single dengue outbreak and provides functions to estimate the final size, the turning point of the epidemic and to conduct a real-time prediction for these parameters using several nonlinear models taking into account model uncertainty using model averaging. Graphical tools for a visualization of the results are also included. 
}
\details{
\tabular{ll}{
Package: \tab DengueRT\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-04-19\cr
License: \tab GPL-3\cr
}

}
\author{
Carlos Sebrango, Lizet Sanchez, Ziv Shkedy, Ewoud De Troyer

Maintainer: Carlos Sebrango <sebrango@uniss.edu.cu>
}
\references{
K. Burnham, D. R. Anderson, Model Selection and Multimodel Inference: A Practical Information-theoretic Approach,
2nd Edition, Springer-Verlag, New York, 2002.\cr
J. MacDougall, Analysis of dose responses Studies: Emax model, in: N. Ting (Ed.), Dose Finding in Drug Development, Statistics for Biology and Health, Springer New York,  pp. 127, 2006.\cr
G. Claeskens, N. L. Hjort, Model selection and model averaging, Cambridge University Press, 2008.\cr
D. Ratkowsky, Handbook of nonlinear regression models, Marcel Dekker, New York, 1990.\cr
F. Richards, A flexible growth function for empirical use, Journal of Experimental Botany 10 (29), pp 290-301, 1959.\cr
Y.H. Hsieh, Temporal trend and regional variability of 2001-2002 multiwave DENV-3 epidemic in Havana City: did Hurricane Michelle contribute to its severity?, Tropical Medicine and International Health, Vol. 18, no. 7, pp 830-838, 2013.\cr
A. Tsoularis, J. Wallace, Analysis of logistic growth models, Mathematical Biosciences, Vol. 179, no. 1, pp 21-55, 2002.\cr
J. Liao, R. Liu, Re-parameterization of five-parameter logistic function, Journal of Chemometrics 23 (5), pp 248-253, 2009.
}
